requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.36.1
h5py==2.10.0
idna==2.10
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
torch==1.8.0
torchvision==0.9.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
pandas==1.0.5
